requirements.txt
# PR0F3550R_LocalAPI - Python Dependencies
# Neural Processing Engine for PR0F3550R-M1NDB0T

# Core Framework
flask>=2.3.0,<3.0.0
numpy>=1.24.0,<2.0.0

# Deep Learning
torch>=2.0.0
torchaudio>=2.0.0

# Audio Processing
librosa>=0.10.0
scipy>=1.10.0

# HTTP Requests (for ElevenLabs TTS integration)
requests>=2.31.0

# Optional: Environment variable management
python-dotenv>=1.0.0

# Optional: Production WSGI server (for deployment)
# gunicorn>=21.0.0

# Optional: WebSocket support (for real-time streaming)
# flask-socketio>=5.3.0
# python-socketio>=5.9.0

# Development Dependencies (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0

